# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Fugue API
  version: 1.0.0
  description: |-
    Authoritative API contract for Fugue.
    Route registration, served OpenAPI artifacts, and generated checks all derive from this file.
servers:
  - url: /
...
paths:
...
  /v1/apps:
    get:
      operationId: listApps
      summary: List Apps
      tags: [Apps]
      x-fugue-handler: handleListApps
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: include_live_status
          required: false
          description: When true, attempt the managed app live status overlay; defaults to false.
          schema:
            type: boolean
        - in: query
          name: include_resource_usage
          required: false
          description: When false, skip the current resource usage overlay.
          schema:
            type: boolean
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      operationId: createApp
      summary: Create App
      tags: [Apps]
      x-fugue-handler: handleCreateApp
      security:
        - BearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /v1/apps/import-image:
    post:
      operationId: importImageApp
      summary: Import Image App
      tags: [Apps]
      x-fugue-handler: handleImportImageApp
      security:
        - BearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportImageRequest"
      responses:
        '202':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportImageResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /v1/apps/{id}:
    get:
      operationId: getApp
      summary: Get App
      tags: [Apps]
      x-fugue-handler: handleGetApp
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    patch:
      operationId: patchApp
      summary: Patch App
      tags: [Apps]
      x-fugue-handler: handlePatchApp
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchAppRequest"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppPatchResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    delete:
      operationId: deleteApp
      summary: Delete App
      tags: [Apps]
      x-fugue-handler: handleDeleteApp
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
        - in: query
          name: force
          required: false
          description: Abort in-flight build or deploy work before deleting the app, and purge the app immediately when no live runtime still needs cleanup.
          schema:
            type: boolean
      responses:
        '202':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppDeleteResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /v1/apps/{id}/images:
    get:
      operationId: getAppImages
      summary: Get App Images
      tags: [Apps]
      x-fugue-handler: handleGetAppImages
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppImageInventoryResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /v1/apps/{id}/rebuild:
    post:
      operationId: rebuildApp
      summary: Rebuild App
      tags: [Apps]
      x-fugue-handler: handleRebuildApp
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RebuildAppRequest"
      responses:
        '202':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppRebuildResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /v1/source-uploads/{id}:
    get:
      operationId: getSourceUpload
      summary: Get Source Upload Metadata
      tags: [Source Uploads]
      x-fugue-handler: handleGetSourceUpload
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceUploadInspectionResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
    RuntimeBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
    EdgeToken:
      type: apiKey
      in: query
      name: token
    DownloadToken:
      type: apiKey
      in: query
      name: download_token
  parameters:
    IdPathParam:
      in: path
      name: id
      required: true
      schema:
        type: string
...
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
    StringMap:
      type: object
      additionalProperties:
        type: string
...
    AppSource:
      type: object
      additionalProperties: false
      properties:
        type:
          type: string
        repo_url:
          type: string
        repo_branch:
          type: string
        repo_auth_token:
          type: string
        image_ref:
          type: string
        resolved_image_ref:
          type: string
        upload_id:
          type: string
        upload_filename:
          type: string
        archive_sha256:
          type: string
        archive_size_bytes:
          type: integer
          format: int64
        source_dir:
          type: string
        build_strategy:
          type: string
        commit_sha:
          type: string
        commit_committed_at:
          type: string
        dockerfile_path:
          type: string
        build_context_dir:
          type: string
        image_name_suffix:
          type: string
        compose_service:
          type: string
        detected_provider:
          type: string
        detected_stack:
          type: string
...
    AppRoute:
      type: object
      additionalProperties: false
      properties:
        hostname:
          type: string
        base_domain:
          type: string
        public_url:
          type: string
        service_port:
          type: integer
          format: int32
    AppInternalService:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
        namespace:
          type: string
        host:
          type: string
        port:
          type: integer
          format: int32
...
    AppFile:
      type: object
      additionalProperties: false
      required: [path]
      properties:
        path:
          type: string
        content:
          type: string
        secret:
          type: boolean
        mode:
          type: integer
          format: int32
...
    AppPersistentStorageSpec:
      type: object
      additionalProperties: false
      properties:
        storage_path:
          type: string
        storage_size:
          type: string
        storage_class_name:
          type: string
        reset_token:
          type: string
        mounts:
          type: array
          items:
            $ref: "#/components/schemas/AppPersistentStorageMount"
...
    AppPostgresSpec:
      type: object
      additionalProperties: false
      properties:
        image:
          type: string
        database:
          type: string
        user:
          type: string
        password:
          type: string
        service_name:
          type: string
        runtime_id:
          type: string
        failover_target_runtime_id:
          type: string
        primary_placement_pending_rebalance:
          type: boolean
        storage_size:
          type: string
        storage_class_name:
          type: string
        instances:
          type: integer
          format: int32
        synchronous_replicas:
          type: integer
          format: int32
        resources:
          $ref: "#/components/schemas/ResourceSpec"
...
    AppStatus:
      type: object
      additionalProperties: false
      required: [phase, current_replicas, updated_at]
      properties:
        phase:
          type: string
        current_runtime_id:
          type: string
        current_replicas:
          type: integer
          format: int32
        current_release_started_at:
          type: string
          format: date-time
        current_release_ready_at:
          type: string
          format: date-time
        last_operation_id:
          type: string
        last_message:
          type: string
        updated_at:
          type: string
          format: date-time
...
    App:
      type: object
      additionalProperties: false
      required: [id, tenant_id, project_id, name, description, spec, status, created_at, updated_at]
      properties:
        id:
          type: string
        tenant_id:
          type: string
        project_id:
          type: string
        name:
          type: string
        description:
          type: string
        source:
          description: Current build source for the app. Prefer build_source for new integrations.
          deprecated: true
          $ref: "#/components/schemas/AppSource"
        origin_source:
          description: Durable source-of-truth used for GitHub sync and source ownership.
          $ref: "#/components/schemas/AppSource"
        build_source:
          description: Last build input applied to the app.
          $ref: "#/components/schemas/AppSource"
        route:
          $ref: "#/components/schemas/AppRoute"
        internal_service:
          $ref: "#/components/schemas/AppInternalService"
        spec:
          $ref: "#/components/schemas/AppSpec"
        status:
          $ref: "#/components/schemas/AppStatus"
        current_resource_usage:
          $ref: "#/components/schemas/ResourceUsage"
        bindings:
          type: array
          items:
            $ref: "#/components/schemas/ServiceBinding"
        backing_services:
          type: array
          items:
            $ref: "#/components/schemas/BackingService"
        tech_stack:
          type: array
          items:
            $ref: "#/components/schemas/AppTechnology"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Operation:
      type: object
      additionalProperties: false
      required: [id, tenant_id, type, status, execution_mode, requested_by_type, requested_by_id, app_id, created_at, updated_at]
      properties:
        id:
          type: string
        tenant_id:
          type: string
        type:
          type: string
        status:
          type: string
        execution_mode:
          type: string
        requested_by_type:
          type: string
        requested_by_id:
          type: string
        app_id:
          type: string
        source_runtime_id:
          type: string
        target_runtime_id:
          type: string
        desired_replicas:
          type: integer
          format: int32
        desired_spec:
          $ref: "#/components/schemas/AppSpec"
        desired_source:
          description: Build input requested by the operation.
          $ref: "#/components/schemas/AppSource"
        desired_origin_source:
          description: Durable source ownership that should persist after the operation completes.
          $ref: "#/components/schemas/AppSource"
        result_message:
          type: string
        manifest_path:
          type: string
        assigned_runtime_id:
          type: string
        error_message:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
...
    RebuildPlan:
      type: object
      additionalProperties: false
      properties:
        source_type:
          type: string
        image_ref:
          type: string
        resolved_image_ref:
          type: string
        branch:
          type: string
        upload_id:
          type: string
        source_dir:
          type: string
        dockerfile_path:
          type: string
        build_context_dir:
          type: string
        build_strategy:
          type: string
        compose_service:
          type: string
    ImportProjectRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
...
    AppImageSummary:
      type: object
      additionalProperties: false
      required:
        [
          version_count,
          current_version_count,
          stale_version_count,
          total_size_bytes,
          current_size_bytes,
          stale_size_bytes,
          reclaimable_size_bytes
        ]
      properties:
        version_count:
          type: integer
          format: int32
        current_version_count:
          type: integer
          format: int32
        stale_version_count:
          type: integer
          format: int32
        total_size_bytes:
          type: integer
          format: int64
        current_size_bytes:
          type: integer
          format: int64
        stale_size_bytes:
          type: integer
          format: int64
        reclaimable_size_bytes:
          type: integer
          format: int64
    AppImageVersion:
      type: object
      additionalProperties: false
      required:
        [image_ref, status, current, delete_supported, redeploy_supported]
      properties:
        image_ref:
          type: string
        runtime_image_ref:
          type: string
        digest:
          type: string
        status:
          type: string
        current:
          type: boolean
        size_bytes:
          type: integer
          format: int64
        reclaimable_size_bytes:
          type: integer
          format: int64
        delete_supported:
          type: boolean
        redeploy_supported:
          type: boolean
        last_deployed_at:
          type: string
          format: date-time
        source:
          $ref: "#/components/schemas/AppSource"
...
    AppImageInventoryResponse:
      type: object
      additionalProperties: false
      required:
        [app_id, registry_configured, reclaim_requires_gc, summary, versions]
      properties:
        app_id:
          type: string
        registry_configured:
          type: boolean
        reclaim_requires_gc:
          type: boolean
        reclaim_note:
          type: string
        summary:
          $ref: "#/components/schemas/AppImageSummary"
        versions:
          type: array
          items:
            $ref: "#/components/schemas/AppImageVersion"
...
    PatchAppRequest:
      type: object
      additionalProperties: false
      properties:
        image_mirror_limit:
          type: integer
          format: int32
        startup_command:
          type: string
        persistent_storage:
          $ref: "#/components/schemas/AppPersistentStorageSpec"
...
    SourceUploadInspection:
      type: object
      additionalProperties: false
      required: [upload, references]
      properties:
        upload:
          $ref: "#/components/schemas/SourceUpload"
        references:
          type: array
          items:
            $ref: "#/components/schemas/SourceUploadReference"
    SourceUploadInspectionResponse:
      type: object
      additionalProperties: false
      required: [source_upload]
      properties:
        source_upload:
          $ref: "#/components/schemas/SourceUploadInspection"
        observed_at:
          type: string
          format: date-time
        error:
          type: string
...
    AppListResponse:
      type: object
      additionalProperties: false
      required: [apps]
      properties:
        apps:
          type: array
          items:
            $ref: "#/components/schemas/App"
    AppResponse:
      type: object
      additionalProperties: false
      required: [app]
      properties:
        app:
          $ref: "#/components/schemas/App"
    AppPatchResponse:
      type: object
      additionalProperties: false
      required: [app, already_current]
      properties:
        app:
          $ref: "#/components/schemas/App"
        already_current:
          type: boolean
        operation:
          $ref: "#/components/schemas/Operation"
...
    ImportImageRequest:
      type: object
      additionalProperties: false
      required: [image_ref]
      properties:
        tenant_id:
          type: string
        project_id:
          type: string
        project:
          $ref: "#/components/schemas/ImportProjectRequest"
        image_ref:
          type: string
        name:
          type: string
        description:
          type: string
        runtime_id:
          type: string
        replicas:
          type: integer
          format: int32
        network_mode:
          $ref: "#/components/schemas/AppNetworkMode"
        service_port:
          type: integer
          format: int32
        env:
          $ref: "#/components/schemas/StringMap"
        config_content:
          type: string
        startup_command:
          type: string
        files:
          type: array
          items:
            $ref: "#/components/schemas/AppFile"
        persistent_storage:
          $ref: "#/components/schemas/AppPersistentStorageSpec"
        postgres:
          $ref: "#/components/schemas/AppPostgresSpec"
    ImportImageResponse:
      type: object
      additionalProperties: false
      required: [app, operation]
      properties:
        app:
          $ref: "#/components/schemas/App"
        operation:
          $ref: "#/components/schemas/Operation"
...
    AppNetworkMode:
      type: string
      description: |
        Omit this field to deploy a public service with a managed route.
        Set `internal` to run the app with a cluster-internal Kubernetes Service
        and readiness checks, but without a public route.
        Set `background` to run the app without a public route, Kubernetes Service,
        or readiness port.
      enum: [background, internal]
...
    RebuildAppRequest:
      type: object
      additionalProperties: false
      properties:
        branch:
          type: string
        image_ref:
          type: string
        source_dir:
          type: string
        dockerfile_path:
          type: string
        build_context_dir:
          type: string
        repo_auth_token:
          type: string
    AppRebuildResponse:
      type: object
      additionalProperties: false
      required: [operation, build]
      properties:
        operation:
          $ref: "#/components/schemas/Operation"
        build:
          $ref: "#/components/schemas/RebuildPlan"
...
    AppDeleteResponse:
      type: object
      additionalProperties: false
      properties:
        operation:
          $ref: "#/components/schemas/Operation"
        already_deleting:
          type: boolean
        deleted:
          type: boolean
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"